Consolidate working with `InjectedLogSite`
