+++ logging/src/commonMain/kotlin/io/spine/logging/InjectedLogSite.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2023, TeamDev. All rights reserved.
+ * Copyright 2025, TeamDev. All rights reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -26,13 +26,112 @@
 
 package io.spine.logging
 
+import com.google.errorprone.annotations.RestrictedApi
+import io.spine.annotation.Internal
+
 /**
- * A [LogSite] that can be manually created and injected with
- * the [LoggingApi.withInjectedLogSite] method.
+ * Creates a log site injected from constants held in a class' constant pool.
+ *
+ * Used for compile-time log site injection, and by the agent.
  */
-public data class InjectedLogSite(
-    override val className: String,
-    override val methodName: String,
-    override val fileName: String?,
-    override val lineNumber: Int,
-) : LogSite()
+@Internal
+@RestrictedApi(
+    explanation =
+        "This function is only used for log-site injection and should not be called directly.",
+    allowlistAnnotations = [LogSiteInjector::class]
+)
+public fun injectedLogSite(
+    internalClassName: String,
+    methodName: String,
+    encodedLineNumber: Int,
+    sourceFileName: String?
+): LogSite = InjectedLogSite(
+    internalClassName,
+    methodName,
+    encodedLineNumber,
+    sourceFileName
+)
+
+/**
+ * A log site implementation used for injected logging information during compile-time or
+ * via agents.
+ *
+ * This class maintains information about the location of a log statement, including the
+ * class name, method name, line number, and source file.
+ *
+ * The class name is stored in the internal JVM format (slash-separated) and converted to
+ * the standard dot-separated format when needed for proper identification.
+ *
+ * @property internalClassName Internal, slash-separated, fully-qualified class name
+ *           (e.g., `"com/example/Foo$Bar"`).
+ * @property method Bare method name without signature information.
+ * @property encodedLineNumber A line number with additional uniqueness
+ *           encoded in the upper 16 bits.
+ * @param sourceFileName The name of the source file containing the log statement.
+ */
+internal class InjectedLogSite(
+    private val internalClassName: String,
+    private val method: String,
+    private val encodedLineNumber: Int,
+    sourceFileName: String?
+) : LogSite() {
+
+    /**
+     * Cached hash code value for optimized lookups.
+     */
+    @Volatile
+    private var hashcode = 0
+
+    /**
+     * Obtains dot-separated class name.
+     *
+     * ## Implementation note
+     *
+     * We have to do the conversion from internal to public class name somewhere, and doing
+     * it earlier could cost work in cases where the log statement is dropped.
+     *
+     * We could cache the result somewhere, but in the default logger backend, this method
+     * is actually only called once anyway when constructing the `LogRecord` instance.
+     */
+    override val className: String
+        get() = internalClassName.replace('/', '.')
+
+    override val methodName: String = method
+
+    /**
+     * Strips additional "uniqueness" information from the upper 16 bits.
+     */
+    @Suppress("MagicNumber")
+    override val lineNumber: Int
+        get() = encodedLineNumber and 0xFFFF
+
+    override val fileName: String? = sourceFileName
+
+    override fun equals(other: Any?): Boolean {
+        if (other is InjectedLogSite) {
+            // Probably not worth optimizing for "this === other" because
+            // all strings should be interned.
+            return method == other.method &&
+                encodedLineNumber == other.encodedLineNumber &&
+                // Check classname last because it isn't cached
+                className == other.className
+        }
+        return false
+    }
+
+    override fun hashCode(): Int {
+        if (hashcode == 0) {
+            // TODO(dbeaumont): Revisit the algorithm when looking at b/22753674.
+            // If the log statement uses metadata, the log site will be used as a key to look up the
+            // current value. In most cases the hashcode is never needed, but in others it may be used
+            // multiple times in different data structures.
+            var temp = 157
+
+            // Don't include classname since it isn't cached. Other fields should be unique enough.
+            temp = 31 * temp + methodName.hashCode()
+            temp = 31 * temp + encodedLineNumber
+            hashcode = temp
+        }
+        return hashcode
+    }
+}
