+++ logging/src/jvmMain/kotlin/io/spine/logging/LogContext.kt
@@ -33,7 +33,6 @@ import io.spine.logging.backend.Metadata
 import io.spine.logging.backend.Platform
 import io.spine.logging.jvm.LogSiteStackTrace
 import io.spine.logging.jvm.context.Tags
-import io.spine.logging.jvm.injectedLogSite
 import io.spine.logging.util.Checks.checkNotNull
 import io.spine.reflect.CallerFinder.stackForCallerOf
 import kotlin.time.DurationUnit
